# Always insert wgmma.fence before MmaOp